requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
joblib==1.2.0
numpy==1.23.4
PyJWT==2.5.0
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
sklearn==0.0
sqlparse==0.4.3
threadpoolctl==3.1.0
whitenoise==6.2.0
